Adding spanwise averaging tool. Sorting slice data using theta coordinate.

+++ tools/fro2vtk/main.cc
@@ -1,6 +1,6 @@
 #include "pptb.h"
 
-using real_t = float;
+using real_t = double;
 
 int main(int argc, char** argv)
 {
